+++ app/adapters/whats_app_adapter/three_sixty_dialog_outbound.rb
@@ -183,18 +183,7 @@ module WhatsAppAdapter
               policy: 'deterministic',
               code: 'de'
             },
-            name: "welcome_message_#{organization.project_name.parameterize.underscore}",
-            components: [
-              {
-                type: 'body',
-                parameters: [
-                  {
-                    type: 'text',
-                    text: organization.project_name
-                  }
-                ]
-              }
-            ]
+            name: "welcome_message_#{organization.project_name.parameterize.underscore}"
           }
         }
       end
