+++ spec/requests/onboarding/whats_app_spec.rb
@@ -193,18 +193,7 @@ RSpec.describe 'Onboarding::Whatsapp' do
                 policy: 'deterministic',
                 code: 'de'
               },
-              name: "welcome_message_#{organization.project_name.parameterize.underscore}",
-              components: [
-                {
-                  type: 'body',
-                  parameters: [
-                    {
-                      type: 'text',
-                      text: organization.project_name
-                    }
-                  ]
-                }
-              ]
+              name: "welcome_message_#{organization.project_name.parameterize.underscore}"
             }
           }
         end
